Refactor form preview into data-driven FormPage with registry
Consolidate FormCertPreview and Form849BPreview into a single FormPage
component driven by a FORM_REGISTRY mapping. Each form entry provides
title, description, generate label, and download filename — eliminating
~370 lines of duplicated code. Also update Form849B title and add
page-break-inside: avoid to narrative section.
